+++ lib/objects/animated_object.dart
@@ -15,8 +15,8 @@ class AnimatedObject extends GameComponent {
 
   @override
   void render(Canvas canvas) {
-    super.render(canvas);
     animation?.getSprite().renderFromVector2Rect(canvas, this.position);
+    super.render(canvas);
   }
 
   @override