+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.3.0
 - new extensions to `GameComponent`.
-- new extensions to `MovementExtensions`.
-- new extensions to `AttackableExtensions`.
+- new extensions to `Movement`.
+- new extensions to `Attackable`.
 - Makes Bonfire more modular. Every kind of behavior has become a mixin.
 
 # 1.2.2