+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - little improvement in `drawDefaultLifeBar`.
 - create mixin `AutomaticRandomMovement`
 - add `onReady` in `BonfireTiledWidget` and `BonfireWidget`
+- add `getComponentDirectionFromMe` in `GameComponentExtensions`
 
 # 1.3.0
 - new extensions to `GameComponent`.