+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# next
+# 1.3.1
 - little improvement in `drawDefaultLifeBar`.
 - create mixin `AutomaticRandomMovement`
 - add `onReady` in `BonfireTiledWidget` and `BonfireWidget`